skip host_observer discovery receiver tests for arm
